requirements.txt
numpy
scipy
scikit-learn
matplotlib
pandas
fastapi
uvicorn
gunicorn==19.9.0
httptools
ccxt
xgboost
google-cloud-bigquery
google-auth
google-cloud-firestore

